move rcreate command to archiver.rcreate
